requirements.txt
asttokens==2.1.0
backcall==0.2.0
black==22.10.0
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
drake==1.10.0
executing==1.2.0
fonttools==4.38.0
importlib-metadata==5.1.0
ipython==8.7.0
kiwisolver==1.4.4
llvmlite==0.39.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
meshcat==0.3.2
mypy-extensions==0.4.3
numba==0.56.4
numpy==1.23.5
packaging==21.3
parso==0.8.3
pathspec==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkg_resources==0.0.0
platformdirs==2.5.4
prompt-toolkit==3.0.33
ptyprocess==0.7.0
pure-eval==0.2.2
pydot==1.4.2
Pygments==2.13.0
pyngrok==5.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==24.0.1
six==1.16.0
stack-data==0.6.2
tomli==2.0.1
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
u-msgpack-python==2.7.2
wcwidth==0.2.5
zipp==3.11.0
